refactor script_with_data_offset macro to crate root
