+++ src/lib.rs
@@ -15,6 +15,7 @@ pub mod memory_client;
 pub mod state;
 pub mod storage;
 pub mod transactor;
+pub mod util;
 
 #[cfg(feature = "profile-any")]
 pub mod profiler;
